+++ examples/current_weather/current_weather.py
@@ -15,7 +15,7 @@ site = conn.get_nearest_forecast_site(51.500728, -0.124626)
 print(site.name)
 
 # Get a forecast for the nearest site
-forecast = conn.get_forecast_for_site(site.id, "3hourly")
+forecast = conn.get_forecast_for_site(site.location_id, "3hourly")
 
 # Get the current timestep using now() and print out some info
 now = forecast.now()